+++ src/main/java/fr/insee/protools/backend/service/platine/pilotage/dto/PlatinePilotageEligibleDto.java
@@ -1,0 +1,9 @@
+package fr.insee.protools.backend.service.platine.pilotage.dto;
+
+import lombok.Data;
+
+@Data
+public class PlatinePilotageEligibleDto {
+
+    private String eligible;
+}